Improve shell scripts with shellcheck

+++ ci/build-lite.sh
@@ -9,25 +9,25 @@ set -x
 set -e
 
 # use RAM disk if possible
-if [ "$CI" == "" ] && [ -d /dev/shm ]; then
+if [ -d /dev/shm ] && mount | grep /dev/shm | grep -v -q noexec; then
     TEMP_BASE=/dev/shm
+elif [ -d /docker-ramdisk ]; then
+    TEMP_BASE=/docker-ramdisk
 else
     TEMP_BASE=/tmp
 fi
 
-BUILD_DIR=$(mktemp -d -p "$TEMP_BASE" AppImageLauncher-build-XXXXXX)
+BUILD_DIR="$(mktemp -d -p "$TEMP_BASE" AppImageLauncher-lite-build-XXXXXX)"
 
-cleanup () {
-    if [ -d "$BUILD_DIR" ]; then
-        rm -rf "$BUILD_DIR"
-    fi
+cleanup() {
+    [[ -d "$BUILD_DIR" ]] && rm -rf "$BUILD_DIR"
 }
 
 trap cleanup EXIT
 
 # store repo root as variable
-REPO_ROOT=$(readlink -f $(dirname "${BASH_SOURCE[0]}")/..)
-OLD_CWD=$(readlink -f .)
+REPO_ROOT="$(readlink -f "$(dirname "${BASH_SOURCE[0]}")"/..)"
+OLD_CWD="$(readlink -f .)"
 
 pushd "$BUILD_DIR"
 
@@ -47,16 +47,16 @@ if [ "$ARCH" == "i386" ]; then
     fi
 fi
 
-cmake "$REPO_ROOT" -DCMAKE_INSTALL_PREFIX=/usr -DCMAKE_BUILD_TYPE=RelWithDebInfo $EXTRA_CMAKE_FLAGS -DCI_BUILD=ON -DBUILD_TESTING=OFF -DBUILD_LITE=ON
+cmake "$REPO_ROOT" -DCMAKE_INSTALL_PREFIX=/usr -DCMAKE_BUILD_TYPE=RelWithDebInfo "${EXTRA_CMAKE_FLAGS[@]}" -DCI_BUILD=ON -DBUILD_TESTING=OFF -DBUILD_LITE=ON
 
 # compile dependencies
-make -j $(nproc) libappimage libappimageupdate libappimageupdate-qt
+make -j "$(nproc)" libappimage libappimageupdate libappimageupdate-qt
 
 # re-run cmake to update paths to dependencies
 cmake .
 
 # build rest
-make -j $(nproc)
+make -j "$(nproc)"
 
 # prepare AppDir
 make install DESTDIR=AppDir
@@ -82,17 +82,18 @@ VERSION="${VERSION}~$(cd "$REPO_ROOT" && git rev-parse --short HEAD)"
 # might seem pointless, but it's necessary to have the version number written inside the AppImage as well, so don't remove
 export VERSION
 
-export OUTPUT=$(echo appimagelauncher-lite-"$VERSION"-"$ARCH".AppImage | tr '~' -)
+OUTPUT="$(echo appimagelauncher-lite-"$VERSION"-"$ARCH".AppImage | tr '~' -)"
+export OUTPUT
 
 export APPIMAGE_EXTRACT_AND_RUN=1
 
 # since we extracted common parts from the installer built into the AppRun script, we have to copy the "library" script
 # before building an AppImage
-install "$REPO_ROOT"/resources/appimagelauncher-lite-installer-common.sh $(readlink -f AppDir/)
+install "$REPO_ROOT"/resources/appimagelauncher-lite-installer-common.sh "$(readlink -f AppDir/)"
 
-./linuxdeploy-"$ARCH".AppImage --plugin qt --appdir $(readlink -f AppDir) --custom-apprun "$REPO_ROOT"/resources/appimagelauncher-lite-AppRun.sh --output appimage \
+./linuxdeploy-"$ARCH".AppImage --plugin qt --appdir "$(readlink -f AppDir)" --custom-apprun "$REPO_ROOT"/resources/appimagelauncher-lite-AppRun.sh --output appimage \
     -d "$REPO_ROOT"/resources/appimagelauncher-lite.desktop \
-    -e $(find AppDir/usr/lib/{,*/}appimagelauncher/remove | head -n1) \
-    -e $(find AppDir/usr/lib/{,*/}appimagelauncher/update | head -n1)
+    -e "$(find AppDir/usr/lib/{,*/}appimagelauncher/remove | head -n1)" \
+    -e "$(find AppDir/usr/lib/{,*/}appimagelauncher/update | head -n1)"
 
 mv "$OUTPUT" "$OLD_CWD"
